split solidbody source files
